add routuing to workflow
