+++ tugas/routers/index.js
@@ -1,0 +1,9 @@
+
+const taskRoutes = require('./task.route');
+
+function main(req, res){
+    taskRoutes(req, res); 
+    // workers routes
+}
+
+module.exports = main
